Enable searching for sequence datasets field, fix bugs.
