+++ tantalus/admin.py
@@ -17,12 +17,7 @@ class SequencingLanesAdmin(admin.ModelAdmin):
 class CurationAdmin(admin.ModelAdmin):
     class Meta:
         model = Curation
-    def save_model(self, request, obj, form, change):
-        super().save_model(request, obj, form, change)
-        print(form.cleaned_data)
-        print(request.user)
-        print(change)
-        print(obj)
+    filter_horizontal = ('sequencedatasets',)
 
 admin.site.register(Tag)
 admin.site.register(Project)
