+++ tantalus/templates/tantalus/curation_list.html
@@ -6,9 +6,11 @@
 {% endblock %}
 
 {% block body %}
-    <div class="container" style="padding:0; margin-bottom:5px;">
+    <div class="container" style="padding:0; margin-bottom:1px;">
         <h4 style="float:left;">Select a Curation to view its details</h4>
+        <a class="btn btn-primary btn-sm" style="float:right; margin-top:5px;" href="{% url 'curation-create' %}">Add a new curation</a>
     </div>
+
     <div class="container" style="background-color:#F5F5F5;">
         {% if curations %}
             {% include "tantalus/datatable/curation_list.html" %}
