+++ tantalus/templates/tantalus/index.html
@@ -60,10 +60,6 @@
                 </tr>
                 <tr class="home-page-row">
                     <td><a href="{% url 'curation-list' %}">Curations</a> <i class="counter">{{curation_count}}</i></td>
-                    <td>
-                        <p style="float:right;">
-                        <a href="{% url 'curation-create' %}"><i class="fa fa-plus" aria-hidden="true" style="color:green;"></i> Add</a></p>
-                    </td>
                 </tr>
                 <tr class="home-page-row">
                     <td><a href="{% url 'external-id-search' %}">Convert External Sample IDs</a></td>
